drivers/cpufreq: round max. scaling frequency to the closest valid frequency
If a new maximum CPU frequency is set either manually by
the user or automatically by the android power HAL that
is not a member of the actually available frequencies,
round the input frequency to the closest valid frequency
included in the available-frequencies-table.

This will for example counter the issue that the CPU will
only run at max. 1036 MHz instead of 1228 MHz, even though
1228 MHz are reported as the current set maximum. Usually,
this may happen when the stock android power HAL is in
place or old init.d scripts for different kernels are
executed.

However, this will not make the 'frequency lock' /
'CPU lock' setting in Trickster MOD unnecessary when over-/
underclocking.
All this does is avoid unnecessary confusion for novice
users.